+++ NEWS.md
@@ -1,6 +1,7 @@
 # assignR news
 
 ## assignR 2.0.0.9000
+* added isoStack function to stack multiple isoscapes in a single data object, and print s3 generic for this class
 
 ## assignR 2.0.0